requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.0
requests>=2.31.0
python-dotenv>=1.0.0
